[RP008] Eliminate trx_mutex use from purge worker threads
innodb_purge_threads seem to conflict still at que_run_threads_low().
All purge threads share the common trx object, and its mutex is used in common path of que0que.cc.
But the common trx is not a real transaction.
Just only for consistency of internal counters, for the worker threads.

This fix stop to use the counters and eliminate the trx->mutex use for the purge worker threads.